requirements.txt
google-generativeai
pyboy
gymnasium
mediapy
einops